+++ musician/__init__.py
@@ -2,7 +2,7 @@
 Package metadata definition.
 """
 
-VERSION = (0, 1, 0, 'final', 0)
+VERSION = (0, 2, 0, 'alpha', 0)
 
 
 def get_version():