+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## master
+- [changed] Include @pangea.org mail addresses (#4).
 
 ## [0.1] - 2020-01-29
 - Login & logout methods using backend as auth method